Moved the PlayerOrderSelect to the components folder
